+++ packages/gov-ui-kit/src/core/components/progress/index.ts
@@ -1,2 +1,2 @@
 export { Progress } from './progress';
-export type { IProgressProps, ProgressSize } from './progress.api';
+export type { IProgressProps, ProgressSize, ProgressVariant } from './progress.api';